requirements.txt
dash==3.3.0
dash-mantine-components==2.4.0
dash-iconify==0.1.2
dash-draggable==0.1.2

Flask==3.0.2
Flask-Session==0.8.0
gunicorn==23.0.0
python-dotenv==1.2.1

msal==1.34.0

pandas==2.3.3
plotly==6.5.0
openpyxl==3.1.5

sqlalchemy==2.0.45
pyodbc==5.3.0

sentry-sdk==2.48.0